move NumberForm and BaseForm to new builtin.inout
